feat: Add route guards

+++ app/src/app/api/link/[id]/route.ts
@@ -11,8 +11,6 @@ export async function GET(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member
-
   try {
     const item = await prisma.link.findUnique({
       where: {
@@ -22,6 +20,15 @@ export async function GET(request: Request, { params }: { params: Params }) {
         tags: true,
       },
     });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) => projectMembership.projectId === item.projectId
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
 
     return NextResponse.json(item);
   } catch (error) {
@@ -34,9 +41,22 @@ export async function PATCH(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member
-
   try {
+    const item = await prisma.link.findUnique({
+      where: {
+        id: params.id,
+      },
+    });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) => projectMembership.projectId === item.projectId
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
+
     const body = await request.json();
 
     await prisma.link.update({
@@ -64,9 +84,22 @@ export async function DELETE(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member
-
   try {
+    const item = await prisma.link.findUnique({
+      where: {
+        id: params.id,
+      },
+    });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) => projectMembership.projectId === item.projectId
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
+
     await prisma.link.delete({
       where: {
         id: params.id,
