+++ app/src/app/api/project/[id]/route.ts
@@ -11,9 +11,24 @@ export async function DELETE(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member and has role 2
-
   try {
+    const item = await prisma.project.findUnique({
+      where: {
+        id: params.id,
+      },
+    });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) =>
+          projectMembership.projectId === item.id &&
+          projectMembership.role === 2
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
+
     await prisma.project.delete({
       where: {
         id: params.id,
