+++ app/src/app/api/tag/[id]/route.ts
@@ -11,14 +11,21 @@ export async function GET(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member
-
   try {
     const item = await prisma.tag.findUnique({
       where: {
         id: params.id,
       },
     });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) => projectMembership.projectId === item.projectId
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
 
     return NextResponse.json(item);
   } catch (error) {
@@ -31,9 +38,22 @@ export async function PATCH(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member
-
   try {
+    const item = await prisma.tag.findUnique({
+      where: {
+        id: params.id,
+      },
+    });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) => projectMembership.projectId === item.projectId
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
+
     const body = await request.json();
 
     await prisma.tag.update({
@@ -57,9 +77,22 @@ export async function DELETE(request: Request, { params }: { params: Params }) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
 
-  // TODO: Check if user is project member
-
   try {
+    const item = await prisma.tag.findUnique({
+      where: {
+        id: params.id,
+      },
+    });
+
+    if (!item) return NextResponse.json({}, { status: 404 });
+
+    if (
+      session.user.projectMemberships.some(
+        (projectMembership) => projectMembership.projectId === item.projectId
+      ) === false
+    )
+      return NextResponse.json({}, { status: 401 });
+
     await prisma.tag.delete({
       where: {
         id: params.id,
